refactor: 피드백 사항 반영
- view가 담당하고 잇는 기능 controller 로 분담
- MVC 패턴 적용
