+++ .eslintignore
@@ -1,3 +1,4 @@
 /node_modules/**
 /src/js/__tests__/**
 /babel.config.js
+*
