--- src/js/ui/utils.js
@@ -1,3 +1,0 @@
-export default function $(id) {
-  return document.getElementById(id);
-}
