+++ src/js/utils/utils.js
@@ -1,3 +1,7 @@
+export function $(id) {
+  return document.getElementById(id);
+}
+
 export function generateRandomNumberRange(MINIMUM, MAXIMUM) {
   return Math.floor(Math.random() * (MAXIMUM - MINIMUM + 1) + MINIMUM);
 }
